refactor: JpaRepository를 서비스 패키지로 이동
- JpaRepository를 구현체가 아닌 인터페이스로 취급하여 서비스 패키지에 넣었습니다. 추후 Repository에는 구현체가 들어갈 예정입니다.
