+++ src/main/java/com/nubble/backend/session/service/UUIDSessionIdGenerator.java
@@ -1,6 +1,5 @@
 package com.nubble.backend.session.service;
 
-import com.nubble.backend.session.respository.SessionRepository;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
