+++ src/test/java/com/nubble/backend/session/service/SessionServiceTest.java
@@ -4,11 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 
 import com.nubble.backend.session.domain.Session;
-import com.nubble.backend.session.respository.SessionRepository;
 import com.nubble.backend.session.service.SessionCommand.SessionCreateCommand;
 import com.nubble.backend.session.service.SessionInfo.SessionCreateInfo;
 import com.nubble.backend.user.domain.User;
-import com.nubble.backend.user.repository.UserRepository;
+import com.nubble.backend.user.service.UserRepository;
 import java.util.Optional;
 import java.util.UUID;
 import org.junit.jupiter.api.Assertions;
